Disconnect GameSessionDecoder channel on unknown packet structure
As the length of the incoming packet is unknown, we can't skip over
the required bytes. If you try to just skip all readable bytes and
not disconnect the channel, it will run into issues if the packet
data is fragmented. Instead we just opt to close the channel immediately
and skip over any of the remaining bytes in the buffer.