Add support to set ActionHandler in ClientPacketStructures
With this change, ActionHandlerMap has been removed. This
allows for a more intuitive way to set a ClientPacket's
handler by setting it on the same builder rather than one
builder for the packet structure and one for the handler
mapping. A KClass<ActionHandler<T>> is used so that we can
inject the handler, which allows for the handler itself to
inject any value it needs