Add `SERVERPROT` + `CLIENTPROT` structure builders
This concept is a rough draft and is subject to change. The Protocol for upstream and downstream packets are created lazily (though can be eager: see TODO @ GamePlatformPacketMaps::eagerInitialize). The goal of this design is to allow these packets to be modified in a "streamline" manner (using DSL in this case as opposed to having a class that implements PacketCodec for each individual packet.) A similar approach was extended to parts of  GameLoginCodec which change on revision upgrades.