Remove `ContentPlugin.enums`
Though the idea of having an easier way to access the values is nice - it is fragile and can be abused easily. `Injector::getInstance` should be avoided at runtime as much as possible. If plugins call this property within methods that will happen at game runtime, then we may get performance hits.