Replace AutoCloseable with `cancel` function
If the class is planned to be used as a singleton - then this feature can be easily abused and have unintended effects. i.e. scope is used in main game dispatcher and another piece of code outside the main module injects the scope and uses the `scope.use` function.

cancel function will throw a scoped cancellation exception to avoid having the supervised completion continuation concurrently remove the child from the scope as its being iterated.