Remove implicit plugin dependencies
Do not want to pollute every plugin with unnecessary dependencies. Opting for a more explicit approach for each module. We _may_ add a user-friendly/implicit way of implementing most commonly-used deps. For example, we can add a commons module with dependencies marked as api and have modules depend on that if they would like all the implicit dependencies that come with it.