Rename `game.plugins` module to `game.scripts`
ContentPlugin name no longer made sense as the plugins were being used for more than game content. Changed the file extension to plugin.kts. This led to confusing package names within the `game.plugin` module and thus the changes in this commit. We _are_ keeping the root project `plugins` module as that is what I'd like them to be known as - at least as far as any content developer is aware.