Initial `user-defined-types` commit
The purpose of this feature is to allow plugins to customize, and/or define, cache type data through flat files. These files should be packed into the cache through a custom task. Packing the data into the cache allows the custom types to be used by the "cache type name generator." Combining these two systems allow the respective type names to be defined in Kotlin constant files and the rscm files (for auto-complete within config files).