Replace LineValidator with LinePathFinder
LinePathFinder.rayCast now returns RayCast, which contains a list of the "travelled" coordinates as well as flags for determening whether the ray cast was successful, partially successful or if it failed. Though we _could_ re-use Route, there is little drawback in creating a new type for a ray cast result and makes the value more explicit (and avoids having to rename Route.waypoints).

LineValidator, LinePathFinder.hasLineOfSight and LinePathFinder.hasLineOfWalk have been marked as deprecated to be removed in the future.