Re-add LineValidator functionality
LineValidator.hasLineOfSight and LineValidator.hasLineOfWalk will be frequently used every game cycle - having to allocate a coordinate list for every call just to use the [RayCast.success] flag is a waste and will lead to a performance hit. The efficient, alternative solution is to re-declare the same code, but return a boolean as opposed to a coordinate list.