Add [GameTestScope.withCollisionState]
This method comes in handy when the test requires control over the collision flags being used in a system. For example, when adding a test for measuring and asserting distance walked in a zone, you wouldn't want collision flags to affect the results. You would opt for a newly-allocated zone with no collision.