Move internal reach strategy related functions
These functions may be useful to the developer without having to use [ReachStrategy.reached] and giving "fake" arguments to use a specific exit strategy. (i.e. if we want to use rectangle strategy we would've had to provide an object shape that represents a rectangle even though we don't care about objects)