Add Kotlin conventions for backing properties
Based on https://kotlinlang.org/docs/properties.html#backing-properties we will add an underscore on backing, mutable properties and add a getter for the non-mutable type. We are mainly doing this to ease the transition to Kotlin's Explicit Backing Fields proposal that will be released soon. (see https://youtrack.jetbrains.com/issue/KT-14663)