Add `droptrigger` support
This is not an rs concept, however it does closely resemble `walktrigger.` This config type will be used to define custom logic when dropping inv objs. For example, when you enter a duel arena combat area the player may have their droptrigger set to a `duel_arena_nodrop` and triggering it will keep resetting it until they are no longer in the area.