Move module `dialogue` -> `player`
The dialogue module relies on `ProtectedAccess` for managing interface dialogue box functions. However, to improve the usability and reduce explicit dependencies, we want to add helper functions in `ProtectedAccess` to start and progress dialogues.

This change introduces a cyclic dependency, but I believe the improved quality of life for plugins justifies this trade-off. Moving the dialogue system into the `player` module keeps related functionality together and simplifies development in the long-run.