Add appearance player-info sync
Not sure how I feel about `ObjTypeList` dependency in `Client.prePlayerCycle` function, however the alternatives are not much better. Though we could make it a bit better by creating a "PlayerCycleContext" of sorts, it would be over-kill as we only require `ObjTypeList` for now. Will probably revise this once the `net` module is reworked in the future.