Improve and update `destroy_note` values
Includes data from wiki, and a "secondary note" field for the destroy notes in differing states. This can include quest progress, or for a different variant of the obj. This value is not packed into the cache, however they serve as a helpful guide when/if filling them out by hand.