Match `opInv` helper functions with protected access interactions
Inv obj interactions are now suspending with protected access, so these helper functions must also be adjusted. The key issue is potential infinite recursion if an `onInvObjN` script calls the corresponding `opInvN` method (and neither introduces a `delay` - time or input based). This would result in recursive calls until [StackOverflowError] is thrown.

We decided to keep these helper functions for convenience. Accordingly, this commit adds a safeguard that tracks how many times `opInv#` is called within the same `ProtectedAccess` scope. If it exceeds a certain threshold, we throw an exception, preventing runaway recursion. However, we may decide to simply discard these functions in the future. (though we'd have to take into account that `InvInteractions` would still be accessible)