Add `results.isNotEmpty` success condition
While this situation would typically occur at a higher level in the codebase (e.g., denying a transaction due to the player being unable to obtain protected access for the respective inventories), it makes sense to only return `success` as true if there were any transactions to begin with. Transactions purposely void of queries should not be considered successful - and ideally, such transactions should not exist at all.