Change `currentMapClock` assignment order in game tick
This change introduces the `updateMapClock` functions, which updates the player's (and npc's) `currentMapClock` to match the global `mapClock`. By doing so, we eliminate the possibility of delays (and queues) being "sped up" by one tick when set by input packets (e.g., delays stemming from `IfButton` clicks).

There is evidence, albeit from a few years ago, that the first "delay" in a queue is not sped up at all (shown in a smithing code screenshot). While this evidence is dated and limited, it is the only concrete information available, and it suggests that delays from input packets should align exactly with the number of ticks specified, rather than being reduced by one.

Even with the lackluster evidence, this change brings a major benefit: it ensures that any developer setting a delay or queuing a task to execute in X ticks can now be certain it will execute in exactly X ticks, regardless of the mechanic or packet triggering it. This guarantees consistent behavior across all systems and eliminates edge cases where delays might behave unpredictably.