Rework `VariableIntMap` and `VariableStringMap` implementations
These classes have been split into entity-specific variants. (e.g., `VarPlayerIntMap` for `Player`) This allows us to place restrictions based on specific needs. For example, for `VarPlayerIntMap` we went ahead and removed the `set` operators functions. When players have their varps modified, we want to send the `Varp` packets instantly. This of course would typically mean handling network-level events in the model layer (business logic). The applied solution was to instead rely on `ProtectedAccess` to provide a `vars` field which acts as a delegate where write-operations end with the associated varp packet being sent to the player's client. This does of course imply a small overhead to construct the delegate the first time it is called during a `ProtectedAccess` scope. (lazy-loaded)

If a plugin script would like to edit a varp/varbit _without_ protected access, they will have to rely on global functions defined in `VarPlayerTypeDelegates`. (i.e., `var Player.foo by intVarp(varps.bar)`) This can also be done with a `ProtectedAccess` instead of a `Player` scope. The main difference is that we may in the future add a "protected varp" concept that exists in rs, where certain varps require protected access in order to be modified.