Replace `withProtectedAccess` call during `countDialog`
Currently, this is only used in countDialog, but it will be applicable to the rest of ResumeP_Input dialogs (e.g., ResumePObjInput).

The previous solution had an issue with cs2-built dialog boxes that could be opened from a modal, leading to both the input dialog (e.g., `countDialog`) and the modal being open simultaneously. In such cases, attempts to "regain" protected access after the suspension point would fail because `Player.isAccessProtected` would incorrectly return true. This happened due to the isBusy check, which verifies if the player is delayed or has any modals open.

This change resolves the issue by comparing the modal opened before the suspension with the one currently active after the suspension. It also ensures that the player is not delayed and that the coroutine they are using matches the one assigned when the `ProtectedAccess` scope was created.