Rework inv syncing process
This rework aligns the inventory syncing process with how `UpdateInvPartial` and `UpdateInvFull` are processed, which we have found to be deferred until later in the tick. An important assumption we've adopted is that `UpdateInvStopTransmit` is sent immediately when `invstoptransmit` is called. Several pieces of evidence support this theory, which guided the implementation.

These changes also address limitations of the previous approach. Now, all inventory scopes are safe to be modified and viewed by any number of players. Previously, this flexibility was limited to "Shared" scope inventories. This improvement is important for working on inventories like CoX shared storage, which are "Temp" scope but can be modified simultaneously by any party member.