Add bank-related query support
Added support for bulk-shift queries, which allow for a range of slots to be shifted either to the left or right of the inventory obj array. This is used quite frequently in bank operations.

Also decided to implement the "compact" operation as a query to separate any modification logic away from the `Inventory` class and stick to using transactions.