Add `transformlink` and `transformtemplate`
Certain objs in the game have a different obj variant when worn and when not worn (i.e., in inv or bank). So similar to `placeholderlink` and `certlink`, these new values are used to switch between the two obj variants. The transaction system now allows for `transform` and `untransform` flags, which work similarly to the `cert` and `uncert` flags.