Add "enriched" cache
This new cache serves as a baseline after initial data from both
plugins and the cache-enrichment system has been applied. Previously,
the "vanilla" cache was used as the base for both the Js5 and Game
caches, but now only the Js5 cache continues to use the vanilla cache,
while the Game cache relies on the enriched cache instead. During
actual packing, both the Js5 and Game cache will source data from the
enriched cache, with encoders ensuring that server-side configs are not
shared with the Js5 cache.