Flip interaction cancellation conditions
The top of the call chain (`shouldCancelInteraction`) remains the same,
however this will make it easier for future changes to each individual
interaction cancellation condition.

With this change, we also deleted (now) unnecessary functions from
`ObjRegistry` and `LocRegistry` that were only there for the sake of
the interaction cancellation function.