Require queue and timer `cycles` arg to be > 0
Since the engine treats the `cycle` arguments of `0` and `1` the same,
it is best to force consistency between the two by removing the ability
to pass in `0` into any "queue" and "timer" functions.

This does not apply to `aitimer` for npcs and controllers as this
concept can be manually "disabled" by passing in a value of `0`.