Add `moveWithMinDelay` functions
This function, along with `walkWithMinDelay` and `runWithMinDelay`,
behaves similarly to their respective variants without the
"WithMinDelay" suffix, but with the added guarantee that the player
will always be delayed for at least 1 cycle, regardless of their
distance to `dest`.

While it's unclear if the official game explicitly supports this
behavior, we do know there are multiple movement procs (e.g.,
`playerwalk1`, `playerwalk2`, `playerwalk3`).

It's possible that one of these variants serves a similar purpose. This
conclusion is based on observed differences in movement behavior
between two scripts: the `Bartender` in Lumbridge and "pickable" crops.