Add mandatory `shape` argument in `findExact`
Unlike `type`, `shape`, and `angle` - `content` is not a valid input
for `findExact`. This means that if none of the initial three args are
passed, then `registry.findExact` will return the first loc it can
find in `coords`.

This can lead to unwanted behavior. When using this refactored
function, the caller should at least be able to specify the associated
`LocShape` they are searching for.