Add `StandardNpcAccess` concept
This system was created to manage npc actions that require access to
their associated `GameCoroutine`. Instead of exposing the coroutine
directly and relying on content to handle it correctly, we introduce
a structured system to manage this access.

Since players already have a similar system (`ProtectedAccess`), the
new system is designed to be consistent so that anyone familiar with
`ProtectedAccess` can easily work with `StandardNpcAccess`.

To maintain consistency with `ProtectedAccess`, we have also introduced
a context system similar to `ProtectedAccessContext`. However, we have
kept it minimal. External dependencies (e.g., NpcTypeList) should be
explicitly passed as arguments in the respective functions rather than
being embedded within the context itself.