Make `EntityList` an abstract class
`EntityList` can be used for more than just managing `Entity`s.
However, since that remains its primary use case, we are keeping the
name for now. It may be renamed in the future if its scope expands.

This change also includes adding a default value to `slotPadding` as
`PlayerList` is (so far) the only implementation that assigns this
value to something other than `0`.