Feat/regions (#30)
* Add `RegionRegistry`

* Add `RegionRepository`

* Add `LocRegistryRegion`

* Add null-zone validation on teleports

* Remove bitpacked "id" for `RemappedRegionLocMap`

Though it's nice to only use the remapped coords for the exact cases
where it's required, this loc id requirement falls apart when the loc
being deleted is not the original loc, but rather a "changed" variant.
(e.g., tree stump being deleted should restore the original tree)

* Clear map flag on invalid `OpLoc`

Although the `isDelayed` condition does not clear the map flag, it
feels slightly off-putting when invalid locs (e.g., locs that went out
of bounds in dynamically-built regions) do not reset the map flag upon
being clicked. This happened even though the player did not route to
them.

* Add `ZoneActivityBitSet`

* Add inactive region removal

* Allocate zone collisions for wander range tests