Add `spawnZone` assignment condition
This assignment is necessary to ensure compatibility with the existing
`add` and `del` logic.

As the class is becoming more verbose, this change also includes
renaming and reorganizing some of the code. While I'm not completely
satisfied with its current state in terms of explicitness and
verbosity, I believe some level of bulk is unavoidable for it to
accomplish everything it needs to.