Add `RegionTemplate` dsl
Replaces the placeholder `RegionZoneBuilder` used to create specific
region templates.

In the future we will add a `RegionDynamicTemplate` which will allow
for dynamic changes to templates, such as copied zones having random
rotations on every region build. (This happens in regions such as
Zulrah where the zones around the main land are randomly rotated)