Revert previous collision flag allocation change
I might have jumped the gun on the previous change in commit
8695ca3549609df2b073091ffb65584d5c39eb6e. I believe that in scenarios
like the one explained, it would be most likely that the player is
teleport straight onto the place that _does_ have collision flags.

In the case of the canoe cutscene, the canoe loc itself does have
collision. This means that only the zone with the canoe in it would be
valid. This makes more sense as one would prefer fully unwalkable zones
from the copied areas to not have any collision flags allocated.