Add `initialLastSlotUsed` arg for `EntityList`
Though subtle, this is a feature for players and npcs. The initial slot
search starts at slot `1`. As of now, there are two cases where this is
required and two where it is not. Controllers may not necessarily
require it, but we _are_ trying to keep them as close to npcs. (as it
only makes sense to)

Any permanent solution will have to wait until we have more use cases
and see which way the "default" behavior should tilt.