Replace `MultiLocEvent` with `base` loc reference
Some multiloc scripts may call `opLoc#` functions, requiring access to
the "base" loc. This change is necessary because the interaction model
verifies the loc's existence each cycle before processing. Technically,
the server is only aware of the "base" loc, as the visual loc depends
on player var values.

Without access to `base`, scripts would need to construct an arbitrary
`LocInfo` with the base loc id, which would be both cumbersome and a
poor design choice. To address this, we've added a `base` loc field to
loc script events. If the loc is not a multiloc, `base` simply
references `loc`.

Since these classes are compact, duplicating data is preferable to
making `base` nullable. The benefits outweigh the cons in this case.