Add `NpcUid`
This uid is explicitly assigned whenever it must be regenerated.
For example, during `npc_add` when a `slotId` is assigned, or
during `npc_del` where the `slotId` is cleared.

This is also done for transmog where the npc's visual type is
changed. It is worth noting that the official game received
support for a transmog command that does _not_ cancel active
npc interactions. The way we will achieve this is by simply
having a new `transmog` function in `StandardNpcAccess` that
does not call `assignUid`.