Rename base `HitModifier`s and `HitProcessor`s
Though the initial thought process was to reduce name pollution for
classes starting with `Player` and `Npc`, we already have a few that
do this, and it has not been an issue; at least not big enough issues
to warrant the old, confusing names.

Also, adds a quick tip in `Npc.queueHit` extension functions for the
`modifier` param.