Add `NpcRepository.despawn` support
Npc deaths would previously re-use "hide" and "reveal" support,
however, there are subtle differences that require respawns to be
handled separately. For example, upon "reveal" (respawn now), the
npc's coords have to be reset to its spawn coords _before_ being
re-added to the zone and applying its collision. Similarly, the
`lastProcessedZone` needs to be set based on this updated coord.