Change `isBusy` condition -> `isAccessProtected`
This subtle but important change ensures that any suspending
action will also "pause" interaction execution. This is crucial
for suspensions like `countDialog`, which display a cs2-built
input chatbox.

The key difference is that, for example, `chatNpc` opens a modal
chatbox, which was already accounted for in `isBusy`. However,
suspend functions like `countDialog` were not previously considered.
With this change, both modal-based interactions and suspending
functions are consistently respected with `isAccessProtected`.