Change Dinh's `Block` attack style
The `Block` attack style requires a non-null value so that it
can be saved and restored correctly when re-equipping the weapon.
However, setting it to `DefensiveMelee` results in unintended
defence roll bonuses from attack style contributions.

To avoid introducing dinh-specific hardcoded checks in the formula,
the block style is assigned to an alternative attack style that is
ignored in combat calculations, since it cannot be used to perform
attacks.