Add default npc melee attack type
We are using category types to store "AttackType" values as
`AttackType` is not a valid `CacheVarLiteral`. This is a more
descriptive way of storing and retrieving the values without
relying on primitive integers.

For now, we only care about the default melee attack type
for npcs. Ranged and magic npcs are expected to have their
own specialized scripts for combat.