Remove `emulateLogInWalkTo` quirk function
This function was originally added to document and replicate a login
edge-case bug. Going forward, this behavior should emerge as a side
effect of the relevant systems, not be explicitly hardcoded.

We will revisit this when working on features like player following
or other mechanics that rely on the player's last known coordinates.