Replace bank `invEquip` call with `opHeld2`
`invEquip` bypasses `opHeld2` scripts assigned to objs, which
is not the intended behavior for bank operations. Equipping objs
from the bank should trigger the same logic as equipping from the
standard inv.

In the future, we should consider having `HeldInteractions.opHeld2`
return a result type. This would let us avoid "cheap" fallback checks
that rely on whether the obj changed in the worn inventory.