Call `p_walk` on manual logout
Credits: https://github.com/2004Scape/Server/issues/1159

This also changes `playerMove` functions to remove the redundancy
check for `coords` and `dest`. This was a remnant of a previous
and stricter `moveSpeed` requirement on players where it should
only be set while the player is actively moving (this is no longer
the case).