Make map builders idempotent
`TypeUpdaterMaps.updateAll` needs to be idempotent as it can be called
multiple times when there is an error that causes the task to re-run.

Previously, the non-idempotent behavior would mask the underlying
error, making debugging more difficult. This change ensures proper
cleanup of map builder singletons after packing, regardless of how
many times the function is called.