Add area music rough draft
Calling this implementation a rough draft as there are various design
choices I am not happy with. However, this will be left as a todo for
the future as it is not worth the time investment for now to make it
"perfect."

There are also various other changes outside of this specific system
that need to be made before cleaning up this implementation (e.g.,
`groupList` api for dbrows requires a lookover).