Replace opheld `clearpendingaction` with `ifclose`
`p_clearpendingaction` is only meant to be called with protected
access. This fixes an issue where opheld packets (`ifbuttonx`) were
able to cancel delayed scripts (e.g., spam clicking mithril seeds to
plant, which should open a dialogue after calling `p_delay`).

This change also removes the `resetFaceEntity` call as we know opheld
is handled in content now, and there is no way for runescript to call
this logic.