Change default exposed `loc` in `oploc#`
`oploc`, `oplocu` and `oploct` handlers previously had `loc` as the
"vis loc" - aka the visual loc taking player varps into account.
Meanwhile, the "base" loc was exposed as `base`.

This has now been reversed. `loc` is now the "base" loc while `vis`
is the "visual loc." This is because most of the time, scripts that
use the loc info tend to require the base loc (e.g., for things such
as repeat `oploc#` calls).